moved and renamed utils_global; cleaned up unused imports
