+++ dataprofiler/profilers/profile_builder.py
@@ -19,7 +19,7 @@ import numpy as np
 import pandas as pd
 from HLL import HyperLogLog
 
-import dataprofiler.generator as generator
+import dataprofiler.utils_global as utils_global
 
 from .. import data_readers, dp_logging
 from ..data_readers.data import Data
@@ -655,7 +655,7 @@ class StructuredColProfiler:
         df_series = df_series.loc[true_sample_list]
         total_na = total_sample_size - len(true_sample_list)
 
-        rng = generator.get_random_number_generator()
+        rng = utils_global.get_random_number_generator()
 
         base_stats = {
             "sample_size": total_sample_size,
