+++ dataprofiler/utils_global.py
